jokoson/requirement.txt
django
djangorestframework
django-filter
djangorestframework-csv
mysql-server
mysqlclient
libmysqlclient-dev